# [SPARK-19008][SQL] Improve performance of Dataset.map by eliminating boxing/unboxing