… sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/plans/logical/object.scala