@cloud-fan 2017-03-07
This is cool! Can you also update the benchmark result in `DatasetBenchmark`?