@kiszk 2017-03-07
[The latest `DatasetBenchmark`](https://github.com/apache/spark/blob/master/sql/core/src/test/scala/org/apache/spark/sql/DatasetBenchmark.scala) has a function `val func = (d: Data) => Data(d.l + 1, d.s)` that this PR cannot be applied to.

Do we add a new benchmark with a function `val func = (d: Data) => Data(d.l + 1)` based on [this suggestion](https://github.com/apache/spark/pull/16391#issuecomment-269361025)?