@cloud-fan 2017-03-08
yea let's add a new case in the benchmark